ignores, helper class for dbs, better pyconsole experience, starting todo items.
Biggest difference here is actually creating an object that I expect
someone to interact with, NamespaceHelper.

Obviously a bad name, but I can improve that over time. For now just
making sure that i provide a way for users to accurately edit data is
important.

Will need a very clear "sync" method so that you never act on a data
set you don't want to.